pyproject.toml
[project]
name = "udm-epic1-synthetic-data"
version = "0.1.0"
description = "UDM Epic 1 — Physics-based Synthetic X-ray Void Data Generator for Universal Detection Model"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "Mohammed Hasan Goni" }]
requires-python = "==3.12.*"
keywords = ["synthetic-data", "x-ray", "void-detection", "semiconductor", "segmentation"]

dependencies = [
    "numpy>=1.24",
    "opencv-python>=4.8",
    "Pillow>=10.0",
    "scipy>=1.11",
    "scikit-image>=0.21",
    "matplotlib>=3.7",
    "tqdm>=4.66",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "typer>=0.9",
    "rich>=13.0",
    "albumentations>=1.3",
    "h5py>=3.9",
    "pandas>=2.0",
    "joblib>=1.3",
    "nbdev>=3.0.12",
    "fastcore>=1.12.31",
    "opencv-python-headless>=4.13.0.92",
    "jupyterlab>=4.5.6",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "timm>=1.0.26",
]

[project.optional-dependencies]
dev = [
    "nbdev>=2.3",
    "jupyter>=1.0",
    "ipykernel>=6.0",
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "black>=23.0",
    "ruff>=0.1",
]
torch = [
    "torch>=2.0",
    "torchvision>=0.15",
    "onnx>=1.14",
    "onnxruntime>=1.16",
]
epic2 = [
    "diffusers>=0.30",
    "accelerate>=0.34",
    "transformers>=4.44",
    "safetensors>=0.4",
]
epic3 = [
    "scikit-image>=0.21",
]
epic4 = [
    "segmentation-models-pytorch>=0.3",
    "scikit-learn>=1.3",
]
epic5 = [
    "scikit-learn>=1.3",
]

[project.scripts]
udm-generate = "udm_epic1.cli:app"
udm-epic2 = "udm_epic2.cli_epic2:app"
udm-epic3 = "udm_epic3.cli_epic3:app"
udm-epic4 = "udm_epic4.cli_epic4:app"
udm-epic5 = "udm_epic5.cli_epic5:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["udm_epic1", "udm_epic2", "udm_epic3", "udm_epic4", "udm_epic5"]

[tool.hatch.build.targets.sdist]
include = ["udm_epic1", "udm_epic2", "udm_epic3", "udm_epic4", "udm_epic5", "nbs", "configs", "scripts", "README.md"]

[tool.nbdev]
lib_path = "udm_epic1_synthetic_data"
nbs_path = "nbs"
doc_path = "docs"
title = "UDM Epic 1 — Synthetic Data Generator"
description = "Physics-based synthetic X-ray void image generation for Universal Detection Model"
author = "Mohammed Hasan Goni"
branch = "main"
repo = "udm-epic1-synthetic-data"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=udm_epic1 --cov=udm_epic2 --cov=udm_epic3 --cov=udm_epic4 --cov=udm_epic5 --cov-report=term-missing"
